+++ src/Controller/SystemsController.php
@@ -51,7 +51,7 @@ class SystemsController extends AppController {
 	 */
 	public function contact_form() {
 		//Checks if the contact form is enabled
-		if(!config('frontend.contact_form')) {
+		if(!config('default.contact_form')) {
 			$this->Session->Error(__d('me_cms', 'Disabled'));
 			return $this->redirect(['_name' => 'homepage']);
 		}
@@ -94,7 +94,7 @@ class SystemsController extends AppController {
 	 */
 	public function offline() {
 		//If the site has not been taken offline
-		if(!config('frontend.offline')) {
+		if(!config('default.offline')) {
 			return $this->redirect(['_name' => 'homepage']);
         }